requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
click==8.1.7
colorama==0.4.6
fastapi==0.115.4
h11==0.14.0
idna==3.10
playsound==1.3.0
pydantic==2.9.2
pydantic_core==2.23.4
pygame==2.6.1
pymodbus==3.7.4
PyYAML==6.0.2
setuptools==75.5.0
sniffio==1.3.1
starlette==0.41.2
typing_extensions==4.12.2
uvicorn==0.32.0
wheel==0.45.0
SQLAlchemy==2.0.36
alembic==1.13.3
ruamel.yaml==0.18.6
argon2-cffi==23.1.0
PyJWT==2.9.0
portalocker==2.10.1

APScheduler==3.10.4
croniter==2.0.5
pytest==8.3.4
httpx==0.27.2
python-dotenv==1.0.0
